--- src/App.css
@@ -1,4 +1,0 @@
-img {
-  width: 120px;
-  height: 120px;
-}
