+++ src/api/data/data.json
@@ -1,122 +1,138 @@
 [
   {
     "id": 1,
-    "phoneImage": "/phoneImages/iphone12.png",
-    "model": "Apple Iphone 12",
-    "brand": "Apple",
-    "releaseDate": 2020,
-    "screenDiagonal": 6.1,
-    "countryOfOrigin": "Китай",
-    "memory": 128,
-    "screenRefreshRate": 60,
-    "nfc": false,
-    "esim": true,
-    "wirelessСharging": true,
-    "price": 81990
+    "name": "Apple",
+    "image": "/phoneImages/iphone12.png",
+    "chars": {
+      "model": "Apple Iphone 12",
+      "releaseDate": 2020,
+      "screenDiagonal": 6.1,
+      "countryOfOrigin": "Китай",
+      "memory": 128,
+      "screenRefreshRate": 60,
+      "nfc": false,
+      "esim": true,
+      "wirelessСharging": true,
+      "price": 81990
+    }
   },
   {
     "id": 2,
-    "phoneImage": "/phoneImages/XiaomiMi11Lite1.png",
-    "model": "Xiaomi Mi 11 Lite",
-    "brand": "Xiaomi",
-    "releaseDate": 2021,
-    "screenDiagonal": 6.55,
-    "countryOfOrigin": "Китай",
-    "memory": 128,
-    "screenRefreshRate": 90,
-    "nfc": true,
-    "esim": true,
-    "wirelessСharging": false,
-    "price": 27490
+    "name": "Xiaomi",
+    "image": "/phoneImages/XiaomiMi11Lite1.png",
+    "chars": {
+      "model": "Xiaomi Mi 11 Lite",
+      "releaseDate": 2021,
+      "screenDiagonal": 6.55,
+      "countryOfOrigin": "Китай",
+      "memory": 128,
+      "screenRefreshRate": 90,
+      "nfc": true,
+      "esim": true,
+      "wirelessСharging": false,
+      "price": 27490
+    }
   },
   {
     "id": 3,
-    "phoneImage": "/phoneImages/A72_1.png",
-    "model": "Samsung Galaxy A72",
-    "brand": "Samsung",
-    "releaseDate": 2021,
-    "screenDiagonal": 6.7,
-    "countryOfOrigin": "Вьетнам",
-    "memory": 128,
-    "screenRefreshRate": 90,
-    "nfc": true,
-    "esim": false,
-    "wirelessСharging": true,
-    "price": 32890
+    "name": "Samsung",
+    "image": "/phoneImages/A72_1.png",
+    "chars": {
+      "model": "Samsung Galaxy A72",
+      "releaseDate": 2021,
+      "screenDiagonal": 6.7,
+      "countryOfOrigin": "Вьетнам",
+      "memory": 128,
+      "screenRefreshRate": 90,
+      "nfc": true,
+      "esim": false,
+      "wirelessСharging": true,
+      "price": 32890
+    }
   },
   {
     "id": 4,
-    "phoneImage": "/phoneImages/SamsungGalaxyS21.png",
-    "model": "Samsung Galaxy S21",
-    "brand": "Samsung",
-    "releaseDate": 2021,
-    "screenDiagonal": 6.2,
-    "countryOfOrigin": "Вьетнам",
-    "memory": 256,
-    "screenRefreshRate": 120,
-    "nfc": true,
-    "esim": true,
-    "wirelessСharging": true,
-    "price": 55299
+    "name": "Samsung",
+    "image": "/phoneImages/SamsungGalaxyS21.png",
+    "chars": {
+      "model": "Samsung Galaxy S21",
+      "releaseDate": 2021,
+      "screenDiagonal": 6.2,
+      "countryOfOrigin": "Вьетнам",
+      "memory": 256,
+      "screenRefreshRate": 120,
+      "nfc": true,
+      "esim": true,
+      "wirelessСharging": true,
+      "price": 55299
+    }
   },
   {
     "id": 5,
-    "phoneImage": "/phoneImages/AppleiPhoneXr.png",
-    "model": "Apple iPhone Xr",
-    "brand": "Apple",
-    "releaseDate": 2022,
-    "screenDiagonal": 6.1,
-    "countryOfOrigin": "Китай",
-    "memory": 256,
-    "screenRefreshRate": 90,
-    "nfc": true,
-    "esim": true,
-    "wirelessСharging": true,
-    "price": 29299
+    "name": "Apple",
+    "image": "/phoneImages/AppleiPhoneXr.png",
+    "chars": {
+      "model": "Apple iPhone Xr",
+      "releaseDate": 2022,
+      "screenDiagonal": 6.1,
+      "countryOfOrigin": "Китай",
+      "memory": 256,
+      "screenRefreshRate": 90,
+      "nfc": true,
+      "esim": true,
+      "wirelessСharging": true,
+      "price": 29299
+    }
   },
   {
     "id": 6,
-    "phoneImage": "/phoneImages/Realme8Pro.png",
-    "model": "Realme 8 Pro",
-    "brand": "Realme",
-    "releaseDate": 2021,
-    "screenDiagonal": 6.4,
-    "countryOfOrigin": "Китай",
-    "memory": 128,
-    "screenRefreshRate": 60,
-    "nfc": true,
-    "esim": true,
-    "wirelessСharging": false,
-    "price": 9000
+    "name": "Realme",
+    "image": "/phoneImages/Realme8Pro.png",
+    "chars": {
+      "model": "Realme 8 Pro",
+      "releaseDate": 2021,
+      "screenDiagonal": 6.4,
+      "countryOfOrigin": "Китай",
+      "memory": 128,
+      "screenRefreshRate": 60,
+      "nfc": true,
+      "esim": true,
+      "wirelessСharging": false,
+      "price": 9000
+    }
   },
   {
     "id": 7,
-    "phoneImage": "/phoneImages/sGalaxyZFlip6.png",
-    "model": "Samsung Galaxy Z Flip6",
-    "brand": "Samsung",
-    "releaseDate": 2024,
-    "screenDiagonal": 6.7,
-    "countryOfOrigin": "Вьетнам",
-    "memory": 256,
-    "screenRefreshRate": 120,
-    "nfc": true,
-    "esim": true,
-    "wirelessСharging": false,
-    "price": 199990
+    "name": "Samsung",
+    "image": "/phoneImages/sGalaxyZFlip6.png",
+    "chars": {
+      "model": "Samsung Galaxy Z Flip6",
+      "releaseDate": 2024,
+      "screenDiagonal": 6.7,
+      "countryOfOrigin": "Вьетнам",
+      "memory": 256,
+      "screenRefreshRate": 120,
+      "nfc": true,
+      "esim": true,
+      "wirelessСharging": false,
+      "price": 199990
+    }
   },
   {
     "id": 8,
-    "phoneImage": "/phoneImages/iphone11.png",
-    "model": "Apple iPhone 11",
-    "brand": "Apple",
-    "releaseDate": 2019,
-    "screenDiagonal": 6.1,
-    "countryOfOrigin": "Индия",
-    "memory": 64,
-    "screenRefreshRate": 60,
-    "nfc": false,
-    "esim": true,
-    "wirelessСharging": false,
-    "price": 37799
+    "name": "Apple",
+    "image": "/phoneImages/iphone11.png",
+    "chars": {
+      "model": "Apple iPhone 11",
+      "releaseDate": 2019,
+      "screenDiagonal": 6.1,
+      "countryOfOrigin": "Индия",
+      "memory": 64,
+      "screenRefreshRate": 60,
+      "nfc": false,
+      "esim": true,
+      "wirelessСharging": false,
+      "price": 37799
+    }
   }
 ]
