+++ src/api/services/getPhones.ts
@@ -1,0 +1,7 @@
+import data from '../data/data.json';
+
+const getPhones = async () => {
+  return await data;
+};
+
+export default getPhones;
