+++ src/index.tsx
@@ -1,9 +1,9 @@
 import React from 'react';
 import { createRoot } from 'react-dom/client';
 import { Provider } from 'react-redux';
-import { store } from './store/store';
+import { store } from './redux/store';
 import App from './App';
-import './index.css';
+import './styles/index.scss';
 
 const container = document.getElementById('root')!;
 const root = createRoot(container);
