+++ src/redux/useAppDispatch/useAppDispatch.ts
@@ -1,0 +1,4 @@
+import { useDispatch } from 'react-redux';
+import type { AppDispatch } from '../store';
+
+export const useAppDispatch = () => useDispatch<AppDispatch>();
