+++ src/redux/useAppSelector/useAppSelector.ts
@@ -1,0 +1,4 @@
+import { TypedUseSelectorHook, useSelector } from 'react-redux';
+import type { RootState } from '../store';
+
+export const useAppSelector: TypedUseSelectorHook<RootState> = useSelector;
