+++ src/styles/index.scss
@@ -1,0 +1,3 @@
+@import 'global.scss';
+@import 'reset.scss';
+@import 'variables.scss';
